Updated pure Numpy feed-forward network
